Change device_model, system_ and app_ versions
By default, the information about the device model,
application version, and operating system was unclear to the user.

Changed the device model to the hostname (platform.uname().node);
set the operating system version to the automatically determined
name and release
(platform.uname().system + ' ' + platform.uname().release),
and set the application version to 0.0.10.