Add an option for a unique length of first part
When editing regular text messages, all parts should not exceed the
standard limit, but when editing media captions, the first part
which is likely to become the caption) should not exceed the
reduced limit for caption length. Therefore, the ability to customize
the length of the first part during splitting, if needed, was added.